requirements.txt
# Project Greenlight - Python Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM Providers
anthropic>=0.18.0
google-generativeai>=0.4.0
google-genai>=1.0.0  # New SDK for Nano Banana Pro (Gemini image generation)

# Image Generation
replicate>=0.22.0

# HTTP Client
httpx>=0.26.0

# Environment
python-dotenv>=1.0.0
